Use default breakpoints values to avoid specific layout config
