+++ projects/storefrontlib/src/recipes/b2c/b2c-layout-config.ts
@@ -1,12 +1,6 @@
 import { LayoutConfig } from '../../layout/config/layout-config';
 
 export const b2cLayoutConfig: LayoutConfig = {
-  breakpoints: {
-    xs: 576,
-    sm: 768,
-    md: 992,
-    lg: 1200,
-  },
   layoutSlots: {
     header: {
       md: {
