+++ projects/storefrontlib/src/shared/components/media/media.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { OccConfig } from '@spartacus/core';
-import { defaultLayoutConfig, LayoutConfig } from '../../../layout';
+import { LayoutConfig } from '../../../layout';
 import { MediaService } from './media.service';
 
 const MockConfig: OccConfig = {
@@ -36,7 +36,8 @@ describe('MediaService', () => {
       providers: [
         MediaService,
         { provide: OccConfig, useValue: MockConfig },
-        { provide: LayoutConfig, useValue: defaultLayoutConfig },
+
+        { provide: LayoutConfig, useValue: {} },
       ],
     });
     mediaService = TestBed.get(MediaService);
